Refactor goHandler some more

+++ source/service/gohandler.go
@@ -38,25 +38,24 @@ var counter int
 
 type GoHandler struct {
 	Prsr             *parser.Parser            // The parser, to tell the GoHandler what the types mean.
-	Modules          map[string]string         // Where the source files are.
 	timeMap          map[string]int            // When the source code was constructed.
-	Plugins          map[string]*plugin.Plugin // Knows where the plugins live after they've been generated.
+	Plugins          *plugin.Plugin // Knows where the plugins live after they've been generated.
 	UserDefinedTypes dtypes.Set[string]        // Set of Pipefish clone/enum/struct types appearing explicitly or implicitly in the sigs of the functions.
-	TypeDeclarations map[string]string         // A string to put the generated source code for declaring things in.
+	TypeDeclarations string         // A string to put the generated source code for declaring things in.
+	output           string
+	sources          dtypes.Set[string]
+	GoImports       []string
 }
 
-func newGoHandler(prsr *parser.Parser) *GoHandler {
+func (cp *Compiler) newGoHandler() {
 
 	gh := GoHandler{
-		Prsr:              prsr,
 		timeMap:           make(map[string]int),
-		Modules:           make(map[string]string),
-		Plugins:     	   make(map[string]*plugin.Plugin),
 		UserDefinedTypes:  make(dtypes.Set[string]),
-		TypeDeclarations:  make(map[string]string),
+		sources:		   make(dtypes.Set[string]),
 	}
 
-	file, err := os.Open(gh.Prsr.Directory + "rsc/go/gotimes.dat")
+	file, err := os.Open(cp.P.Directory + "rsc/go/gotimes.dat")
 	if err != nil {
 		panic("Can't open file 'rsc/go/gotimes.dat'.")
 	}
@@ -73,46 +72,48 @@ func newGoHandler(prsr *parser.Parser) *GoHandler {
 		gh.timeMap[lines[2*i]] = time
 	}
 
-	return &gh
+	cp.goHandler = &gh
 }
 
 // This is called directly from the initializer to shove the pure Go blocks into the code.
-func (gh *GoHandler) addPureGoBlock(source, code string) {
-	gh.Modules[source] = gh.Modules[source] + "\n" + code[:len(code)-2] + "\n\n"
+func (gh *GoHandler) addPureGoBlock(code string) {
+	gh.output = gh.output + "\n" + code[:len(code)-2] + "\n\n"
 }
 
 func (cp *Compiler) getGoFunctions() {
-	for source := range cp.goHandler.Modules {
-		cp.transitivelyCloseTypes()
-		cp.goHandler.TypeDeclarations[source] = cp.generateDeclarations()
-		if cp.P.ErrorsExist() {
-			return
-		}
+	
+	if len(cp.goHandler.sources) == 0 {
+		return
 	}
-	cp.goHandler.buildGoModules()
+
+	cp.transitivelyCloseTypes()
+	cp.goHandler.TypeDeclarations = cp.generateDeclarations()
+	if cp.P.ErrorsExist() {
+		return
+	}
+	
+	cp.buildGoModules()
 	if cp.P.ErrorsExist() {
 		return
 	}
 	newGoConverter := make([](func(t uint32, v any) any), len(cp.Vm.concreteTypeInfo))
 	copy(newGoConverter, cp.Vm.goConverter)
-	for source := range cp.goHandler.Modules {
-		functionConverterSymbol, _ := cp.goHandler.Plugins[source].Lookup("PIPEFISH_FUNCTION_CONVERTER")
-		functionConverter := *functionConverterSymbol.(*map[string](func(t uint32, v any) any))
-		maps.Copy(functionConverter, BUILTIN_FUNCTION_CONVERTER)
-		for typeName, constructor := range functionConverter {
-			typeNumber := cp.concreteTypeNow(typeName)
-			newGoConverter[typeNumber] = constructor
-		}
-		cp.Vm.goConverter = newGoConverter
-		valueConverterSymbol, _ := cp.goHandler.Plugins[source].Lookup("PIPEFISH_VALUE_CONVERTER")
-		valueConverter := *valueConverterSymbol.(*map[string]any)
-		maps.Copy(valueConverter, BUILTIN_VALUE_CONVERTER)
-		for typeName, goValue := range valueConverter {
-			cp.Vm.goToPipefishTypes[reflect.TypeOf(goValue).Elem()] = cp.concreteTypeNow(typeName)
-		}
+	
+	functionConverterSymbol, _ := cp.goHandler.Plugins.Lookup("PIPEFISH_FUNCTION_CONVERTER")
+	functionConverter := *functionConverterSymbol.(*map[string](func(t uint32, v any) any))
+	maps.Copy(functionConverter, BUILTIN_FUNCTION_CONVERTER)
+	for typeName, constructor := range functionConverter {
+		typeNumber := cp.concreteTypeNow(typeName)
+		newGoConverter[typeNumber] = constructor
+	}
+	cp.Vm.goConverter = newGoConverter
+	valueConverterSymbol, _ := cp.goHandler.Plugins.Lookup("PIPEFISH_VALUE_CONVERTER")
+	valueConverter := *valueConverterSymbol.(*map[string]any)
+	maps.Copy(valueConverter, BUILTIN_VALUE_CONVERTER)
+	for typeName, goValue := range valueConverter {
+		cp.Vm.goToPipefishTypes[reflect.TypeOf(goValue).Elem()] = cp.concreteTypeNow(typeName)
 	}
 	
-	// TODO --- see if this plays nicely with function sharing and modules or if it needs more work.
 	if cp.P.NamespacePath == "" {
 		for k, v := range cp.P.Common.Functions {
 			if v.Body.GetToken().Type == token.GOCODE {
@@ -129,7 +130,7 @@ func (cp *Compiler) getGoFunctions() {
 			}
 		}
 	}
-	cp.goHandler.recordGoTimes()
+	cp.recordGoTimes()
 }
 
 var BUILTIN_FUNCTION_CONVERTER = map[string](func(t uint32, v any) any){
@@ -182,59 +183,82 @@ func (cp *Compiler) transitivelyCloseTypes() {
 	}
 }
 
-func (gh *GoHandler) buildGoModules() {
+func (cp *Compiler) buildGoModules() {
 
-	for source, functionBodies := range gh.Modules {
+	// We find if we need to rebuild the Go.
 
-		var modifiedTime int64
-		f, err := os.Stat(MakeFilepath(source, gh.Prsr.Directory))
+	// The purpose of putting timestamps on the .so files is not that we ever read the timestamps
+	// from the filenames (we look either at the OS metadata of the file or at the 'gotimes' file),
+	// but simply because you can't re-use the names of .so files in the same Go runtime and since
+	// we're looking up times anyway this is a reasonable way to achieve that.
+
+	needsRebuild := false
+	var lastChange int
+	var maxLastChange int 
+	var maxModifiedTime int64
+
+	for source := range cp.goHandler.sources {
+
+		f, err := os.Stat(MakeFilepath(source, cp.P.Directory))
+		if err != nil {
+			cp.Throw("go/file", token.Token{Source: "linking Golang"}, err.Error())
+			needsRebuild = false
+			break
+		}
+		modifiedTime := f.ModTime().UnixMilli()
+		if modifiedTime > maxModifiedTime {
+			maxModifiedTime = modifiedTime
+		}
+		lastChange, ok := cp.goHandler.timeMap[source]
+		if lastChange > maxLastChange {
+			maxLastChange = lastChange
+		}
+		if !ok || modifiedTime != int64(lastChange) {
+			needsRebuild = true 
+			break
+		}
+	}
+
+	if !needsRebuild {
+		soFile := cp.P.Directory + "rsc/go/" + text.Flatten(cp.ScriptFilepath) + ".so"
+		plugins, err := plugin.Open(soFile)
 		if err == nil {
-			modifiedTime = f.ModTime().UnixMilli()
+			cp.goHandler.Plugins = plugins
+			return
 		}
-
-		lastChange, ok := gh.timeMap[source]
-		if ok {
-			if modifiedTime == int64(lastChange) {
-				soFile := gh.Prsr.Directory + "rsc/go/" + text.Flatten(source) + "_" + strconv.Itoa(lastChange) + ".so"
-				gh.Plugins[source], err = plugin.Open(soFile)
-				if err == nil {
-					continue
-				}
-				println("Error building/using .so file")
-				println("Error was", err.Error())
-				panic("That's all folks.")
-			}
-		}
+		println("Error building/using .so file")
+		println("Error was", err.Error())
+		panic("That's all folks.")
+	} else {
 
 		preface := "package main\n\n"
 
-		if len(gh.Prsr.GoImports[source]) > 0 {
+		if len(cp.goHandler.GoImports) > 0 {
 			preface = preface + "import (\n"
-			for _, v := range gh.Prsr.GoImports[source] {
+			for _, v := range cp.goHandler.GoImports {
 				preface = preface + "    \"" + v + "\"\n"
 			}
 			preface = preface + ")\n\n"
 		}
 
-		// You can't reuse the names of shared object files.
-		counter++
-		soFile := filepath.Join(gh.Prsr.Directory, filepath.FromSlash("rsc/go/"+text.Flatten(source)+"_"+strconv.Itoa(int(modifiedTime))+".so"))
+		counter++ // The number of the gocode_<counter>.go source file we're going to write.
+		soFile := filepath.Join(cp.P.Directory, filepath.FromSlash("rsc/go/"+ text.Flatten(cp.ScriptFilepath) + "_" + strconv.Itoa(int(maxModifiedTime)) + ".so"))
 		if lastChange != 0 {
-			os.Remove(filepath.Join(gh.Prsr.Directory, filepath.FromSlash("rsc/go/"+text.Flatten(source)+"_"+strconv.Itoa(int(lastChange))+".so")))
+			os.Remove(filepath.Join(cp.P.Directory, filepath.FromSlash("rsc/go/"+ text.Flatten(cp.ScriptFilepath) + "_" + strconv.Itoa(maxLastChange) + ".so")))
 		}
-		goFile := filepath.Join(gh.Prsr.Directory, "gocode_"+strconv.Itoa(counter)+".go")
+		goFile := filepath.Join(cp.P.Directory, "gocode_"+strconv.Itoa(counter)+".go")
 		file, _ := os.Create(goFile)
-		file.WriteString(preface + gh.TypeDeclarations[source] + functionBodies)
+		file.WriteString(preface + cp.goHandler.TypeDeclarations + cp.goHandler.output)
 		file.Close()
 		cmd := exec.Command("go", "build", "-buildmode=plugin", "-o", soFile, goFile) // Version to use running from terminal.
 		// cmd := exec.Command("go", "build", "-gcflags=all=-N -l", "-buildmode=plugin", "-o", soFile, goFile) // Version to use with debugger.
 		output, err := cmd.Output()
 		if err != nil {
-			gh.Prsr.Throw("golang/build", &token.Token{}, err.Error()+": "+string(output))
+			cp.P.Throw("golang/build", &token.Token{}, err.Error()+": "+string(output))
 		}
-		gh.Plugins[source], err = plugin.Open(soFile)
+		cp.goHandler.Plugins, err = plugin.Open(soFile)
 		if err != nil {
-			gh.Prsr.Throw("golang/open", &token.Token{}, err.Error())
+			cp.P.Throw("golang/open", &token.Token{}, err.Error())
 		}
 		if err == nil || strings.Contains(err.Error(), "plugin was built with a different version of package") {
 			os.Remove("gocode_" + strconv.Itoa(counter) + ".go")
@@ -242,27 +266,27 @@ func (gh *GoHandler) buildGoModules() {
 	}
 }
 
-func (gh *GoHandler) recordGoTimes() {
+func (cp *Compiler) recordGoTimes() {
 
 	// We add the newly compiled modules to the list of times.
 
-	for k := range gh.Modules {
-		filepath := MakeFilepath(k, gh.Prsr.Directory)
+	for k := range cp.goHandler.sources {
+		filepath := MakeFilepath(k, cp.P.Directory)
 		file, err := os.Stat(filepath)
 		if err != nil {
 			panic("Gohandler cleanup: " + err.Error())
 		}
 		modifiedTime := file.ModTime().UnixMilli()
-		gh.timeMap[k] = int(modifiedTime)
+		cp.goHandler.timeMap[k] = int(modifiedTime)
 	}
 	// And then write out the list of times to the .dat file.
 
-	f, err := os.Create(gh.Prsr.Directory + "rsc/go/gotimes.dat")
+	f, err := os.Create(cp.P.Directory + "rsc/go/gotimes.dat")
 	if err != nil {
 		panic("Can't create file rsc/go/gotimes.dat")
 	}
 	defer f.Close()
-	for k, v := range gh.timeMap {
+	for k, v := range cp.goHandler.timeMap {
 		f.WriteString(k + "\n")
 		f.WriteString(strconv.Itoa(v) + "\n")
 	}
@@ -270,6 +294,6 @@ func (gh *GoHandler) recordGoTimes() {
 
 func (gh *GoHandler) getFn(fnName string, tok *token.Token) reflect.Value {
 	name := text.Capitalize(fnName)
-	fn, _ := gh.Plugins[tok.Source].Lookup(name)
+	fn, _ := gh.Plugins.Lookup(name)
 	return reflect.ValueOf(fn)
 }
